refactor: move build_pvc_job_payload from conftest.py to utils.py
